+++ embassy-stm32/src/timer/pwm_input.rs
@@ -1,6 +1,6 @@
 //! PWM Input driver.
 
-use super::low_level::{CountingMode, InputCaptureMode, InputTISelection, SlaveMode, Timer, TriggerSource};
+use super::low_level::{CountingMode, InputCaptureMode, InputTISelection, NoIrq, SlaveMode, Timer, TriggerSource};
 use super::{Ch1, Ch2, Channel, GeneralInstance4Channel, TimerPin};
 use crate::gpio::{AfType, Pull};
 use crate::time::Hertz;
@@ -13,7 +13,7 @@ use crate::Peri;
 /// Double check your chips reference manual
 pub struct PwmInput<'d, T: GeneralInstance4Channel> {
     channel: Channel,
-    inner: Timer<'d, T>,
+    inner: Timer<'d, T, NoIrq>,
 }
 
 impl<'d, T: GeneralInstance4Channel> PwmInput<'d, T> {